# Conflicts:
#	src/jdk.jshell/share/classes/jdk/jshell/SourceCodeAnalysisImpl.java

+++ test/hotspot/jtreg/runtime/cds/appcds/jcmd/JCmdTestFileSafety.java
@@ -27,17 +27,19 @@
  * @bug 8265465 8267075
  * @summary Test jcmd to dump static shared archive.
  * @requires vm.cds
- * @library /test/lib /test/hotspot/jtreg/runtime/cds/appcds
+ * @library /test/lib /test/hotspot/jtreg/runtime/cds/appcds /test/hotspot/jtreg/runtime/cds/appcds/test-classes
  * @modules jdk.jcmd/sun.tools.common:+open
- * @compile ../test-classes/Hello.java JCmdTestDumpBase.java
- * @build sun.hotspot.WhiteBox
- * @build JCmdTestLingeredApp JCmdTestFileSafety
- * @run driver jdk.test.lib.helpers.ClassFileInstaller sun.hotspot.WhiteBox
+ * @build jdk.test.lib.apps.LingeredApp jdk.test.whitebox.WhiteBox Hello
+ *        JCmdTestDumpBase JCmdTestLingeredApp JCmdTestFileSafety
+ * @run driver jdk.test.lib.helpers.ClassFileInstaller jdk.test.whitebox.WhiteBox
  * @run main/othervm/timeout=480 -Xbootclasspath/a:. -XX:+UnlockDiagnosticVMOptions -XX:+WhiteBoxAPI JCmdTestFileSafety
  */
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.attribute.FileTime;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import jdk.test.lib.cds.CDSTestUtils;
 import jdk.test.lib.apps.LingeredApp;
 import jdk.test.lib.Platform;
@@ -61,6 +63,13 @@ public class JCmdTestFileSafety extends JCmdTestDumpBase {
         }
     }
 
+    private static void removeFile(String fileName) throws Exception {
+        File file = new File(fileName);
+        if (file.exists()) {
+            file.delete();
+        }
+    }
+
     static void test() throws Exception {
         buildJars();
 
@@ -75,6 +84,8 @@ public class JCmdTestFileSafety extends JCmdTestDumpBase {
         }
         outputDirFile.setWritable(true);
         String localFileName = subDir + File.separator + "MyStaticDump.jsa";
+        removeFile(localFileName);
+
         setIsStatic(true/*static*/);
         // Set target dir not writable, do static dump
         print2ln(test_count++ + " Set target dir not writable, do static dump");
@@ -82,10 +93,16 @@ public class JCmdTestFileSafety extends JCmdTestDumpBase {
         app = createLingeredApp("-cp", allJars);
         pid = app.getPid();
         test(localFileName, pid, noBoot,  EXPECT_PASS);
+        checkFileExistence(localFileName, true/*exist*/);
+        FileTime ft1 = Files.getLastModifiedTime(Paths.get(localFileName));
         outputDirFile.setWritable(false);
         test(localFileName, pid, noBoot,  EXPECT_FAIL);
+        FileTime ft2 = Files.getLastModifiedTime(Paths.get(localFileName));
+        if (!ft2.equals(ft1)) {
+            throw new RuntimeException("Archive file " + localFileName + " should not be updated");
+        }
+        removeFile(localFileName);
         outputDirFile.setWritable(true);
-        checkFileExistence(localFileName, true/*exist*/);
 
         // Illegal character in file name
         localFileName = "mystatic:.jsa";
@@ -103,18 +120,17 @@ public class JCmdTestFileSafety extends JCmdTestDumpBase {
         pid = app.getPid();
         localFileName = subDir + File.separator + "MyDynamicDump.jsa";
         test(localFileName, pid, noBoot,  EXPECT_PASS);
-        app.stopApp();
-        // cannot dynamically dump twice, restart
-        app = createLingeredApp("-cp", allJars, "-XX:+RecordDynamicDumpInfo");
-        pid = app.getPid();
+        checkFileExistence(localFileName, true/*exist*/);
+        ft1 = Files.getLastModifiedTime(Paths.get(localFileName));
         outputDirFile.setWritable(false);
         test(localFileName, pid, noBoot,  EXPECT_FAIL);
-        outputDirFile.setWritable(true);
+        ft2 = Files.getLastModifiedTime(Paths.get(localFileName));
+        if (!ft2.equals(ft1)) {
+            throw new RuntimeException("Archive file " + localFileName + " should not be updated");
+        }
         app.stopApp();
-        // MyDynamicDump.jsa should exist
-        checkFileExistence(localFileName, true);
-        File rmFile = new File(localFileName);
-        rmFile.delete();
+        removeFile(localFileName);
+        outputDirFile.setWritable(true);
         outputDirFile.delete();
     }
